+++ src/components/produtos.jsx
@@ -1,153 +1,155 @@
-import React from "react";
-import styled from "styled-components";
+// import React from "react";
+// import styled from "styled-components";
 
-const Produto = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  text-align: center;
-  gap: 15px;
-  border: solid 1px black;
-  padding-bottom: 15px;
-`;
-const MainContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-`;
+// const Produto = styled.div`
+//   display: flex;
+//   flex-direction: column;
+//   justify-content: center;
+//   align-items: center;
+//   text-align: center;
+//   gap: 15px;
+//   border: solid 1px black;
+//   padding-bottom: 15px;
+// `;
+// const MainContainer = styled.div`
+//   display: flex;
+//   flex-direction: column;
+// `;
 
-const Imagem = styled.img`
-  width: 400px;
-  height: 400px;
-`;
+// const Imagem = styled.img`
+//   width: 400px;
+//   height: 400px;
+// `;
 
-const Corpo = styled.div`
-  flex-wrap: wrap;
-  justify-content: center;
-  gap: 100px;
-`;
+// const Corpo = styled.div`
+//   flex-wrap: wrap;
+//   justify-content: center;
+//   gap: 100px;
+// `;
 
-class Produtos extends React.Component {
-  state = {
-    produtos: [
-      {
-        id: 1,
-        name: "Aome do Produto",
-        value: 10000.0, //valor do produto
-        imagem: "https://picsum.photos/400/400?a",
-      },
-      {
-        id: 2,
-        name: "Bome do Produto 2",
-        value: 50.0, //valor do produto
-        imagem: "https://picsum.photos/400/400?b",
-      },
-      {
-        id: 3,
-        name: "Come do Produto 3",
-        value: 120.0, //valor do produto
-        imagem: "https://picsum.photos/400/400?c",
-      },
-      {
-        id: 4,
-        name: "Dome do Produto 4",
-        value: 300.0, //valor do produto
-        imagem: "https://picsum.photos/400/400?d",
-      },
-      {
-        id: 5,
-        name: "Eome do Produto 5",
-        value: 1000.0, //valor do produto
-        imagem: "https://picsum.photos/400/400?e",
-      },
-      {
-        id: 6,
-        name: "Fome do Produto 6",
-        value: 2500.0, //valor do produto
-        imagem: "https://picsum.photos/400/400?f",
-      },
-    ],
+// class Produtos extends React.Component {
+//   state = {
+//     produtos: [
+//       {
+//         id: 1,
+//         name: "Aome do Produto",
+//         value: 10000.0, //valor do produto
+//         imagem: "https://picsum.photos/400/400?a",
+//       },
+//       {
+//         id: 2,
+//         name: "Bome do Produto 2",
+//         value: 50.0, //valor do produto
+//         imagem: "https://picsum.photos/400/400?b",
+//       },
+//       {
+//         id: 3,
+//         name: "Come do Produto 3",
+//         value: 120.0, //valor do produto
+//         imagem: "https://picsum.photos/400/400?c",
+//       },
+//       {
+//         id: 4,
+//         name: "Dome do Produto 4",
+//         value: 300.0, //valor do produto
+//         imagem: "https://picsum.photos/400/400?d",
+//       },
+//       {
+//         id: 5,
+//         name: "Eome do Produto 5",
+//         value: 1000.0, //valor do produto
+//         imagem: "https://picsum.photos/400/400?e",
+//       },
+//       {
+//         id: 6,
+//         name: "Fome do Produto 6",
+//         value: 2500.0, //valor do produto
+//         imagem: "https://picsum.photos/400/400?f",
+//       },
+//     ],
 
-    selecionandoParametro: "name",
-    ordenandoProduto: 1,
-  };
+//     selecionandoParametro: "name",
+//     ordenandoProduto: 1,
+//   };
 
-  updateSelecionandoParametro = (event) => {
-    this.setState({
-      selecionandoParametro: event.target.value,
-    });
-  };
+//   updateSelecionandoParametro = (event) => {
+//     this.setState({
+//       selecionandoParametro: event.target.value,
+//     });
+//   };
 
-  updateOrdenandoProduto = (event) => {
-    this.setState({
-      ordenandoProduto: event.target.value,
-    });
-  };
+//   updateOrdenandoProduto = (event) => {
+//     this.setState({
+//       ordenandoProduto: event.target.value,
+//     });
+//   };
 
-  render() {
-    const qtdeDeArray = (
-      <div>Quantidade de Produtos: {this.state.produtos.length}</div>
-    );
+//   render() {
+//     const qtdeDeArray = (
+//       <div>Quantidade de Produtos: {this.state.produtos.length}</div>
+//     );
 
-    const ordenandoProduto = (
-      <div>
-        <label for="sort">Ordenação</label>
-        <select
-          name="sort"
-          value={this.state.selecionandoParametro}
-          onChange={this.updateSelecionandoParametro}
-        >
-          <option value="name">Título</option>
-          <option value="value">Preço</option>
-        </select>
+//     const ordenandoProduto = (
+//       <div>
+//         <label for="sort">Ordenação</label>
+//         <select
+//           name="sort"
+//           value={this.state.selecionandoParametro}
+//           onChange={this.updateSelecionandoParametro}
+//         >
+//           <option value="name">Título</option>
+//           <option value="value">Preço</option>
+//         </select>
 
-        <select
-          name="ordenando"
-          value={this.state.ordenandoProduto}
-          onChange={this.updateOrdenandoProduto}
-        >
-          <option value={1}>Crescente</option>
-          <option value={-1}>Decrescente</option>
-        </select>
-      </div>
-    );
+//         <select
+//           name="ordenando"
+//           value={this.state.ordenandoProduto}
+//           onChange={this.updateOrdenandoProduto}
+//         >
+//           <option value={1}>Crescente</option>
+//           <option value={-1}>Decrescente</option>
+//         </select>
+//       </div>
+//     );
+//     const copiaDosProdutos = [...this.state.produtos];
+//     copiaDosProdutos.sort((primeiroProduto, segundoProduto) => {
+//       switch (this.state.selecionandoParametro) {
+//         case "name":
+//           return (
+//             this.state.ordenandoProduto *
+//             primeiroProduto.name.localeCompare(segundoProduto.name)
+//           );
+//         default:
+//           return (
+//             this.state.ordenandoProduto *
+//             (primeiroProduto.value - segundoProduto.value)
+//           );
+//       }
+//     });
 
-    this.state.produtos.sort((primeiroProduto, segundoProduto) => {
-      switch (this.state.selecionandoParametro) {
-        case "name":
-          return (
-            this.state.ordenandoProduto *
-            primeiroProduto.name.localeCompare(segundoProduto.name)
-          );
-        default:
-          return (
-            this.state.ordenandoProduto *
-            (primeiroProduto.value - segundoProduto.value)
-          );
-      }
-    });
+//     const listagemDeProdutos = copiaDosProdutos.map((produto) => {
+//       return (
+//         <Produto>
+//           <Imagem src={produto.imagem} />
+//           <p>{produto.name}</p>
+//           <p>R$: {produto.value}</p>
+//           <button>Adicionar ao Carrinho</button>
+//         </Produto>
+//       );
+//     });
+//     return (
+//       <MainContainer>
+//         <header>
+//           {qtdeDeArray}
+//           {ordenandoProduto}
+//         </header>
 
-    const listagemDeProdutos = this.state.produtos.map((produto) => {
-      return (
-        <Produto>
-          <Imagem src={produto.imagem} />
-          <p>{produto.name}</p>
-          <p>R$: {produto.value}</p>
-          <button>Adicionar ao Carrinho</button>
-        </Produto>
-      );
-    });
-    return (
-      <MainContainer>
-        <header>
-          {qtdeDeArray}
-          {ordenandoProduto}
-        </header>
+//         <Corpo>
+//           {listagemDeProdutos}
+//         </Corpo>
+//       </MainContainer>
+//     );
+//   }
+// }
 
-        <Corpo>{listagemDeProdutos}</Corpo>
-      </MainContainer>
-    );
-  }
-}
-
-export default Produtos;
+// export default Produtos;
